refactor: split member and item endpoints
