+++ src/services/item/tasks/base-public-item-task.ts
@@ -1,9 +1,9 @@
 // global
-import { FastifyLoggerInstance } from 'fastify';
-import { Actor, DatabaseTransactionHandler, ItemService } from 'graasp';
-import { Task, TaskStatus } from 'graasp';
+import { FastifyLoggerInstance } from "fastify";
+import { Actor, DatabaseTransactionHandler, ItemService } from "graasp";
+import { Task, TaskStatus } from "graasp";
 // local
-import { PublicItemService } from '../db-service';
+import { PublicItemService } from "../db-service";
 
 export abstract class BasePublicItemTask<R> implements Task<Actor, R> {
   protected publicItemService: PublicItemService;
@@ -20,12 +20,19 @@ export abstract class BasePublicItemTask<R> implements Task<Actor, R> {
     this.actor = actor;
     this.publicItemService = publicItemService;
     this.itemService = itemService;
-    this.status = 'NEW';
+    this.status = "NEW";
   }
 
   abstract get name(): string;
-  get result(): R { return this._result; }
-  get message(): string { return this._message; }
+  get result(): R {
+    return this._result;
+  }
+  get message(): string {
+    return this._message;
+  }
 
-  abstract run(handler: DatabaseTransactionHandler, log: FastifyLoggerInstance): Promise<void | BasePublicItemTask<R>[]>;
+  abstract run(
+    handler: DatabaseTransactionHandler,
+    log: FastifyLoggerInstance
+  ): Promise<void | BasePublicItemTask<R>[]>;
 }
